Linting Remove re-export private type in public interface
