+++ packages/claude-sdk/src/index.ts
@@ -2,8 +2,8 @@ import { AnthropicClient } from './private/AnthropicClient';
 import { ApprovalCoordinator } from './private/ApprovalCoordinator';
 import { AnthropicAuth } from './private/Client/Auth/AnthropicAuth';
 import type { AuthCredentials } from './private/Client/Auth/types';
-import { ControlChannel } from './private/ControlChannel';
 import type { IPublisher, ISubscriber } from './private/ControlChannel';
+import { ControlChannel } from './private/ControlChannel';
 import { Conversation } from './private/Conversation';
 import { calculateCost } from './private/pricing';
 import { QueryRunner } from './private/QueryRunner';
