+++ packages/claude-sdk/src/private/types.ts
@@ -1,3 +1,5 @@
+import type { ContentBlock } from '../public/types';
+
 export type ApprovalResponse = {
   approved: boolean;
   reason?: string;
@@ -8,21 +10,6 @@ export type ToolUseResult = {
   name: string;
   input: Record<string, unknown>;
 };
-
-export type ServerToolResultBlock = {
-  type: 'web_search_tool_result' | 'web_fetch_tool_result' | 'code_execution_tool_result' | 'bash_code_execution_tool_result' | 'text_editor_code_execution_tool_result' | 'tool_search_tool_result' | 'mcp_tool_result';
-  toolUseId: string;
-  content: unknown;
-};
-
-export type ContentBlock =
-  | { type: 'thinking'; thinking: string; signature: string }
-  | { type: 'text'; text: string }
-  | { type: 'tool_use'; id: string; name: string; input: Record<string, unknown> }
-  | { type: 'compaction'; content: string }
-  | { type: 'server_tool_use'; id: string; name: string; input: Record<string, unknown> }
-  | ServerToolResultBlock
-  | { type: 'redacted_thinking'; data: string };
 
 export type MessageUsage = {
   inputTokens: number;
